feat: refactor màn simple tree
