+++ src/components/TreePage/SimpleView/GridConceptRelation/index.js
@@ -1,0 +1,1 @@
+export { default } from './GridConceptRelation';
