+++ src/components/TreePage/SimpleView/ListExample/index.js
@@ -1,0 +1,1 @@
+export { default } from './ListExample';
