+++ src/components/TreePage/SimpleView/ListExampleRelation/index.js
@@ -1,0 +1,1 @@
+export { default } from './ListExampleRelation';
