+++ src/components/TreePage/SimpleView/SimpleView.js
@@ -1,269 +1,71 @@
-import { East as ArrowRightIcon, Edit, North as ArrowUpIcon } from '@mui/icons-material';
-import { Box, Grid, IconButton, ListItemButton, ListItemText, Paper, Tooltip, Typography } from '@mui/material';
+import { North as ArrowUpIcon } from '@mui/icons-material';
+import { Box } from '@mui/material';
+import { useQuery } from '@tanstack/react-query';
 import { memo, useEffect, useState } from 'react';
-import Markdown from '~/components/BaseComponent/Markdown';
-import EditExampleDialog from '~/components/Example/EditExampleDialog';
-import { getDisplayExample, stripHtmlExceptHighlight } from '~/utils/common/utils';
-import { stylePaper, textEllipsisText } from '~/utils/style/muiCustomStyle';
+import GridConceptRelation from '~/components/TreePage/SimpleView/GridConceptRelation';
+import ListExample from '~/components/TreePage/SimpleView/ListExample';
+import ListExampleRelation from '~/components/TreePage/SimpleView/ListExampleRelation';
+import { getTree } from '~/services/treeService';
 
-function SimpleView({ treeData, listExample, onClickConcept, onClickRelationType, reloadShowTree }) {
-    const [openDialog, setOpenDialog] = useState(false);
-    const [selectedExampleId, setSelectedExampleId] = useState(null);
+const ArrowUpBox = () => {
+    return (
+        <Box
+            sx={{
+                display: 'flex',
+                justifyContent: 'center',
+                alignItems: 'center',
+                my: 1,
+            }}
+        >
+            <ArrowUpIcon color="primary" />
+        </Box>
+    );
+};
+
+function SimpleView({ rootConcept, setRootConcept }) {
     const [selectedExampleLinkId, setSelectedExampleLinkId] = useState(null);
 
-    const handleClickExample = (exampleId) => {
-        setSelectedExampleId(exampleId);
-    };
+    useEffect(() => {
+        setSelectedExampleLinkId(null);
+    }, [rootConcept]);
 
-    const handleDbClickExample = (exampleId) => {
-        setSelectedExampleId(exampleId);
-        setOpenDialog(true);
-    };
-
-    const handleAfterModifyExample = () => {
-        reloadShowTree();
-        onClickRelationType(selectedExampleLinkId);
-    };
+    const { data: treeData } = useQuery({
+        queryKey: ['getTree', rootConcept?.ConceptId],
+        queryFn: async () => {
+            const res = await getTree(rootConcept?.ConceptId);
+            return res.data.Data;
+        },
+        enabled: !!rootConcept?.ConceptId,
+    });
 
     const handleClickConcept = (concept) => {
-        onClickConcept(concept);
+        setRootConcept(concept);
     };
 
-    const handleClickRelationType = (exampleLinkId) => {
-        setSelectedExampleLinkId(exampleLinkId);
-        onClickRelationType(exampleLinkId);
-    };
-
-    useEffect(() => {
-        setSelectedExampleId(null);
-        setSelectedExampleLinkId(null);
-    }, [treeData]);
+    if (!rootConcept || !treeData) {
+        return <></>;
+    }
 
     return (
         <>
-            {openDialog && (
-                <EditExampleDialog
-                    open={openDialog}
-                    onClose={() => setOpenDialog(false)}
-                    exampleId={selectedExampleId}
-                    handleAfter={handleAfterModifyExample}
-                />
-            )}
-            <Paper sx={{ ...stylePaper, p: 2, mt: 2 }}>
-                <Typography>
-                    <Typography component="span" color="primary" sx={{ fontWeight: '500', mb: 1 }}>
-                        List examples:
-                    </Typography>
-                    {listExample.length === 0 ? ' No data' : ''}
-                </Typography>
-                {/* {listExample?.length > 0 && (
-                    <div style={{ display: 'flex', marginBottom: '4px' }}>
-                        <Typography
-                            variant="caption"
-                            color="text.secondary"
-                            sx={{ display: 'flex', alignItems: 'flex-start' }}
-                        >
-                            <Info fontSize="small" color="text.secondary" sx={{ mr: 0.2, pb: 0.2 }} />
-                            Double-click to view/edit/delete example
-                        </Typography>
-                    </div>
-                )} */}
-                <div style={{ maxHeight: 400, overflow: 'auto' }}>
-                    {listExample?.map((x, index) => (
-                        <ListItemButton
-                            key={index}
-                            sx={{ px: 2 }}
-                            onClick={() => handleClickExample(x.ExampleId)}
-                            onDoubleClick={() => handleDbClickExample(x.ExampleId)}
-                            selected={x.ExampleId === selectedExampleId}
-                        >
-                            <ListItemText>
-                                <Markdown
-                                    children={`${index + 1} - ${getDisplayExample(
-                                        stripHtmlExceptHighlight(x.ExampleHtml),
-                                    )}`}
-                                />
-                            </ListItemText>
-                            <Tooltip title="View/Edit/Delete">
-                                <IconButton onClick={() => handleDbClickExample(x.ExampleId)}>
-                                    <Edit fontSize="small" />
-                                </IconButton>
-                            </Tooltip>
-                        </ListItemButton>
-                    ))}
-                </div>
-            </Paper>
-            <Box
-                sx={{
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                    my: 1,
-                }}
-            >
-                <ArrowUpIcon color="primary" />
-            </Box>
-            <Paper sx={{ ...stylePaper, p: 2, width: '100%', overflow: 'auto' }}>
-                <Box
-                    sx={{
-                        display: 'flex',
-                        flexDirection: 'row',
-                        alignItems: 'center',
-                        flexWrap: 'nowrap',
-                        // justifyContent: 'center',
-                        gap: '24px',
-
-                        '&:before, &:after': {
-                            content: '""',
-                            margin: 'auto',
-                        },
-                    }}
-                >
-                    {treeData.ListCountExampleLink?.map((x, i) => (
-                        <ListItemButton
-                            sx={{
-                                flexDirection: 'column',
-                                flexShrink: 0,
-                                px: 2,
-                            }}
-                            key={i}
-                            onClick={() => handleClickRelationType(x.ExampleLinkId)}
-                            selected={x.ExampleLinkId === selectedExampleLinkId}
-                            disabled={x.Count === 0}
-                        >
-                            <Typography color="primary" sx={{ whiteSpace: 'nowrap' }}>
-                                {x.ExampleLinkName}
-                            </Typography>
-                            <Typography color="primary" variant="h6">
-                                {x.Count}
-                            </Typography>
-                        </ListItemButton>
-                    ))}
-                </Box>
-            </Paper>
-            <Box
-                sx={{
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                    my: 1,
-                }}
-            >
-                <ArrowUpIcon color="primary" />
-            </Box>
-            <Grid container spacing={1}>
-                <Grid item xs={12} sm={12} md={4} order={{ md: 2, sm: 1, xs: 1 }}>
-                    <Box
-                        sx={{
-                            display: 'flex',
-                            justifyContent: 'center',
-                            alignItems: 'center',
-                            my: 1,
-                        }}
-                    >
-                        <Box sx={{ display: { md: 'block', sm: 'none', xs: 'none' } }}>
-                            <ArrowRightIcon color="primary" />
-                        </Box>
-                        <Paper
-                            sx={{
-                                ...stylePaper,
-                                display: 'flex',
-                                justifyContent: 'center',
-                                alignItems: 'center',
-                                width: 200,
-                                height: 50,
-                                p: 2,
-                                m: '0 auto',
-                                bgcolor: 'primary.main',
-                                color: 'primary.contrastText',
-                                overflow: 'hidden',
-                                borderRadius: '40px',
-                            }}
-                            id="paper1"
-                        >
-                            <Tooltip title={treeData.Concept?.Title}>
-                                <Typography variant="h6" sx={textEllipsisText}>
-                                    {treeData.Concept?.Title}
-                                </Typography>
-                            </Tooltip>
-                        </Paper>
-                        <Box sx={{ display: { md: 'block', sm: 'none', xs: 'none' } }}>
-                            <ArrowRightIcon color="primary" />
-                        </Box>
-                    </Box>
-                </Grid>
-                <Grid item xs={12} sm={6} md={4} order={{ md: 1, sm: 1, xs: 2 }}>
-                    <Paper
-                        sx={{
-                            ...stylePaper,
-                            width: 280,
-                            maxHeight: 300,
-                            p: 2,
-                            m: '0 auto',
-                        }}
-                        id="paper2"
-                    >
-                        <Typography color="primary" fontWeight="500" pl={1}>
-                            Children
-                        </Typography>
-                        <Box>
-                            {treeData.ListChildren?.map((x) => (
-                                <Tooltip title={`${x.Title} (${x.ConceptLinkName})`} key={x.ConceptId}>
-                                    <ListItemButton sx={{ px: 1, py: 0.5 }} onClick={() => handleClickConcept(x)}>
-                                        <ListItemText
-                                            sx={{ maxWidth: '50%' }}
-                                            primaryTypographyProps={{ style: textEllipsisText }}
-                                        >
-                                            {x.Title}
-                                        </ListItemText>
-                                        <ListItemText
-                                            sx={{ textAlign: 'right', maxWidth: '50%' }}
-                                            primaryTypographyProps={{ style: textEllipsisText }}
-                                        >
-                                            {x.ConceptLinkName}
-                                        </ListItemText>
-                                    </ListItemButton>
-                                </Tooltip>
-                            ))}
-                        </Box>
-                    </Paper>
-                </Grid>
-                <Grid item xs={12} sm={6} md={4} order={{ md: 3, sm: 3, xs: 3 }}>
-                    <Paper
-                        sx={{
-                            ...stylePaper,
-                            width: 280,
-                            maxHeight: 300,
-                            p: 2,
-                            m: '0 auto',
-                        }}
-                    >
-                        <Typography color="primary" fontWeight="500">
-                            Parent
-                        </Typography>
-                        <Box>
-                            {treeData.ListParent?.map((x) => (
-                                <Tooltip title={`${x.Title} (${x.ConceptLinkName})`} key={x.ConceptId}>
-                                    <ListItemButton sx={{ px: 1, py: 0.5 }} onClick={() => handleClickConcept(x)}>
-                                        <ListItemText
-                                            sx={{ maxWidth: '50%' }}
-                                            primaryTypographyProps={{ style: textEllipsisText }}
-                                        >
-                                            {x.ConceptLinkName}
-                                        </ListItemText>
-                                        <ListItemText
-                                            sx={{ textAlign: 'right', maxWidth: '50%' }}
-                                            primaryTypographyProps={{ style: textEllipsisText }}
-                                        >
-                                            {x.Title}
-                                        </ListItemText>
-                                    </ListItemButton>
-                                </Tooltip>
-                            ))}
-                        </Box>
-                    </Paper>
-                </Grid>
-            </Grid>
+            <ListExample
+                rootConceptId={treeData?.Concept?.ConceptId}
+                exampleLinkId={selectedExampleLinkId}
+                setExampleLinkId={setSelectedExampleLinkId}
+            />
+            <ArrowUpBox />
+            <ListExampleRelation
+                listCountExampleLink={treeData?.ListCountExampleLink}
+                selectedId={selectedExampleLinkId}
+                setSelectedId={setSelectedExampleLinkId}
+            />
+            <ArrowUpBox />
+            <GridConceptRelation
+                rootConcept={treeData?.Concept}
+                listChildren={treeData?.ListChildren || []}
+                listParent={treeData?.ListParent || []}
+                onClickConcept={handleClickConcept}
+            />
         </>
     );
 }
