+++ src/pages/TreePage/TreePage.js
@@ -3,13 +3,9 @@ import styles from './TreePage.module.scss';
 import classNames from 'classnames/bind';
 import { Box, Button, FormControlLabel, Switch, Typography } from '@mui/material';
 import useLocalStorage from '~/hooks/useLocalStorage';
-import { useMutation } from '@tanstack/react-query';
 import ConceptAutocomplete from '~/components/TreePage/ConceptAutocomplete';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { toast } from 'react-toastify';
-import { getLinkedExampleByRelationshipType, getTree } from '~/services/treeService';
-import Loading from '~/components/Loading';
-import { Enum } from '~/utils/common/enumeration';
 import SimpleView from '~/components/TreePage/SimpleView';
 import searchImg from '~/assets/images/search.svg';
 import { useLocation } from 'react-router-dom';
@@ -22,8 +18,12 @@ function TreePage() {
 
     const [selectedConcept, setSelectedConcept] = useState(null);
     const [rootConcept, setRootConcept] = useState(null);
-    const [treeData, setTreeData] = useState({});
-    const [listExample, setListExample] = useState([]);
+
+    useEffect(() => {
+        if (rootConcept) {
+            setSelectedConcept(rootConcept);
+        }
+    }, [rootConcept]);
 
     // =============================================================================
 
@@ -49,44 +49,6 @@ function TreePage() {
 
     // =============================================================================
 
-    const { mutate: handleShowTree, isLoading: isLoadingShowTree } = useMutation(
-        async () => {
-            const res = await getTree(rootConcept?.ConceptId);
-            return res.data;
-        },
-        {
-            onSuccess: (data) => {
-                if (data?.Status === Enum.ServiceResultStatus.Success) {
-                    setTreeData(data.Data);
-                } else if (data?.Status === Enum.ServiceResultStatus.Fail) {
-                    setTreeData({});
-                    toast.error(data.Message || 'Show failed');
-                } else {
-                    setTreeData({});
-                    toast.error('Show failed');
-                }
-            },
-        },
-    );
-
-    const { mutate: getExampleByRelationType } = useMutation(
-        async (exampleLinkId) => {
-            const res = await getLinkedExampleByRelationshipType(rootConcept?.ConceptId, exampleLinkId);
-            return res.data;
-        },
-        {
-            onSuccess: (data) => {
-                if (data?.Status === Enum.ServiceResultStatus.Success) {
-                    setListExample(data.Data);
-                } else if (data?.Status === Enum.ServiceResultStatus.Fail) {
-                    setListExample([]);
-                } else {
-                    setListExample([]);
-                }
-            },
-        },
-    );
-
     const handleClickShowTree = () => {
         if (selectedConcept) {
             setRootConcept(selectedConcept);
@@ -94,30 +56,6 @@ function TreePage() {
             toast.warning('Concept is required');
         }
     };
-
-    const handleClickConceptItem = useCallback((concept) => {
-        setSelectedConcept(concept);
-        setRootConcept(concept);
-    }, []);
-
-    const handleClickRelationType = useCallback(
-        (exampleLinkId) => {
-            getExampleByRelationType(exampleLinkId);
-        },
-        [getExampleByRelationType],
-    );
-
-    useEffect(() => {
-        if (rootConcept) {
-            setListExample([]);
-            setSelectedConcept(rootConcept);
-            handleShowTree();
-        }
-    }, [rootConcept, handleShowTree]);
-
-    const reloadShowTree = useCallback(() => {
-        handleShowTree();
-    }, [handleShowTree]);
 
     return (
         <div className={cx('wrapper')}>
@@ -141,8 +79,6 @@ function TreePage() {
                     }}
                 />
             </Box>
-
-            {isLoadingShowTree && <Loading />}
             <Box className={cx('main-wrapper')} sx={{ mt: 2 }}>
                 <Box className={cx('search-wrapper')}>
                     <ConceptAutocomplete
@@ -165,15 +101,7 @@ function TreePage() {
                         <img src={searchImg} alt="search" className={cx('img-no-data')} />
                     </Box>
                 )}
-                {!fullMode && !!rootConcept && (
-                    <SimpleView
-                        treeData={treeData}
-                        listExample={listExample}
-                        onClickConcept={handleClickConceptItem}
-                        onClickRelationType={handleClickRelationType}
-                        reloadShowTree={reloadShowTree}
-                    />
-                )}
+                {!fullMode && !!rootConcept && <SimpleView rootConcept={rootConcept} setRootConcept={setRootConcept} />}
                 {fullMode && !!rootConcept && <FullView rootConcept={rootConcept} setRootConcept={setRootConcept} />}
             </Box>
         </div>
